+++ src/biip/gs1/__init__.py
@@ -28,13 +28,13 @@ Example:
     format=GtinFormat.GTIN_13, prefix=GS1Prefix(value='703', usage='GS1
     Norway'), payload='703206980498', check_digit=8, packaging_level=None),
     date=None, decimal=None, money=None)
-    >>> msg.element_strings[1]
+    >>> msg.get(data_title='BEST BY')
     GS1ElementString(ai=GS1ApplicationIdentifier(ai='15', description='Best
     before date (YYMMDD)', data_title='BEST BEFORE or BEST BY',
     fnc1_required=False, format='N2+N6'), value='210526',
     pattern_groups=['210526'], gtin=None, date=datetime.date(2021, 5, 26),
     decimal=None, money=None)
-    >>> msg.element_strings[2]
+    >>> msg.get(ai="10")
     GS1ElementString(ai=GS1ApplicationIdentifier(ai='10', description='Batch
     or lot number', data_title='BATCH/LOT', fnc1_required=True,
     format='N2+X..20'), value='0329', pattern_groups=['0329'], gtin=None,